ai/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tensorflow>=2.14.0
scikit-learn>=1.3.0
pandas>=2.1.0
matplotlib>=3.7.0
torch>=2.1.0
torchvision>=0.16.0
python-jose>=3.3.0
passlib>=1.7.4 